+++ templates/workspace_services/sql/terraform/locals.tf
@@ -1,5 +1,5 @@
 locals {
-  short_service_id               = substr(var.id, -4, -1)
+  short_service_id               = substr(var.tre_resource_id, -4, -1)
   short_workspace_id             = substr(var.workspace_id, -4, -1)
   workspace_resource_name_suffix = "${var.tre_id}-ws-${local.short_workspace_id}"
   service_resource_name_suffix   = "${var.tre_id}-ws-${local.short_workspace_id}-svc-${local.short_service_id}"
@@ -10,5 +10,10 @@ locals {
     "GP | 5GB 4vCores" = { value = "GP_Gen5_4" },
     "GP | 5GB 6vCores" = { value = "GP_Gen5_6" },
     "GP | 5GB 8vCores" = { value = "GP_Gen5_8" }
+  }
+  workspace_service_tags = {
+    tre_id                   = var.tre_id
+    tre_workspace_id         = var.workspace_id
+    tre_workspace_service_id = var.tre_resource_id
   }
 }