+++ templates/workspace_services/sql/terraform/main.tf
@@ -3,7 +3,7 @@ terraform {
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = "=3.5.0"
+      version = "3.18.0"
     }
     random = {
       source  = "hashicorp/random"