+++ templates/workspace_services/sql/terraform/mysql.tf
@@ -22,6 +22,7 @@ resource "azurerm_mysql_server" "mysql" {
   public_network_access_enabled     = false
   ssl_enforcement_enabled           = true
   ssl_minimal_tls_version_enforced  = "TLS1_2"
+  tags                              = local.workspace_service_tags
 
   lifecycle { ignore_changes = [tags] }
 }
@@ -39,6 +40,7 @@ resource "azurerm_private_endpoint" "mysql_private_endpoint" {
   location            = data.azurerm_resource_group.ws.location
   resource_group_name = data.azurerm_resource_group.ws.name
   subnet_id           = data.azurerm_subnet.services.id
+  tags                = local.workspace_service_tags
 
   private_service_connection {
     private_connection_resource_id = azurerm_mysql_server.mysql.id