+++ templates/workspace_services/sql/terraform/variables.tf
@@ -1,6 +1,6 @@
 variable "workspace_id" {}
 variable "tre_id" {}
-variable "id" {}
+variable "tre_resource_id" {}
 variable "sql_sku" {}
 variable "db_name" {}
 variable "storage_mb" {